fix: Handle identifiers wrapped in "
Queries containing " inside them were not handled correct. The reason
was that when checking all the regex "--sql was checked before """--sql
but the first would match the latter. So when wrapping python stings
with tripple quptes it would think it was "--sql and end the string on
the first " it found which could have been wrapped around a column name.
e.g.
```python
sql = """--sql
    SELECT "group" FROM some_table;
"""
```
This would only match the `SELECT ` and then stop because it encountered
a `"`.